Migrate history observations to internal IDs
Replace natural-key identifiers with internal bigint IDs in history.observations (connector_id, timeseries_id), add created_at, and remove moved_at. Convert numeric fields to double precision across schemas (observations, timeseries last_value, RPC aggregates). Replace the old multi-column index with a BRIN on observed_at. Change upsert behavior to conflict on (connector_id, timeseries_id, observed_at) and perform conditional updates only when value/status differ to avoid unnecessary writes. Add since_ts filtering and include_status option to several RPCs (uk_aq_la_hex_rpc, uk_aq_pcon_hex_rpc, uk_aq_timeseries_rpc) and update grants/overloads accordingly. Fix RLS table name for breathelondon checkpoints and update related docs and README (add planning egress/DB-size guidance).

+++ schemas/main_db/uk_aq_core_schema.sql
@@ -361,7 +361,7 @@ create table if not exists timeseries (
   category_id bigint references categories(id),
   first_value_at timestamptz,
   last_value_at timestamptz,
-  last_value numeric,
+  last_value double precision,
   extras jsonb,
   rendering_hints jsonb,
   status_intervals jsonb,
@@ -416,7 +416,7 @@ create table if not exists observations (
   connector_id bigint not null references connectors(id) on delete cascade,
   timeseries_id bigint references timeseries(id) on delete cascade,
   observed_at timestamptz not null,
-  value numeric,
+  value double precision,
   status text,
   created_at timestamptz default now(),
   primary key (connector_id, timeseries_id, observed_at)
